don't lock ejs and eco versions
Currently backbone-on-rails locks eco and ejs to versions ~> 1.0.0, 
and as far as I can tell there isn't a reason for this: an eco or ejs 
API change could break an app using backbone-on-rails, but since
backbone-on-rails doesn't directly use either gem (right?), it 
couldn't break backbone-on-rails itself. We should leave it up to
users to determine which version of eco or ejs they want.

Long story short: ejs 1.0.0 has some significant bugs which were fixed
a while ago, but we can't use the newer versions alongside 
backbone-on-rails unless we unlock the version here.